requirements.txt
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
openpyxl==3.1.2
reportlab==4.0.9
Pillow==10.2.0
django-otp==1.4.0
qrcode==7.4.2
requests==2.31.0
django-extensions==3.2.3
black==24.3.0
flake8==7.0.0
django-filter==25.2
